requirements.txt
beautifulsoup4==4.12.3
polars==1.20.0
Requests==2.32.3
tqdm==4.67.1
pyarrow==19.0.0
python-dotenv==1.0.1
geopy==2.4.1
cachetools==5.5.1
fastapi==0.115.6
uvicorn[standard]==0.34.0
jinja2==3.1.5
aiosqlite==0.20.0
httpx==0.28.1
pytest==8.3.4
pytest-asyncio==0.25.0
